Tranforms of Mapped Items fixed to correctly position geometry
git-tfs-id: [https://tfs.codeplex.com:443/tfs/TFS13]$/xBIM/XbimFramework/Dev/ModelServer;C35004

+++ Xbim.Ifc.Extensions/Axis2Placement3DExtensions.cs
@@ -82,12 +82,12 @@ namespace Xbim.Ifc2x3.Extensions
                 xa.Normalize();
                 XbimVector3D ya = XbimVector3D.CrossProduct(za, xa);
                 ya.Normalize();
-                return new XbimMatrix3D(xa.X, xa.Y, xa.Z, 0, ya.X, ya.Y, ya.Z, 0, za.X, za.Y, za.Z, 0, (float)axis3.Location.X,
-                                    (float)axis3.Location.Y, (float)axis3.Location.Z, 1);
+                return new XbimMatrix3D(xa.X, xa.Y, xa.Z, 0, ya.X, ya.Y, ya.Z, 0, za.X, za.Y, za.Z, 0, axis3.Location.X,
+                                    axis3.Location.Y, axis3.Location.Z, 1);
             }
             else
-                return new XbimMatrix3D(1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, (float)axis3.Location.X, (float)axis3.Location.Y,
-                                    (float)axis3.Location.Z, 1);
+                return new XbimMatrix3D(1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, axis3.Location.X, axis3.Location.Y,
+                                    axis3.Location.Z, 1);
 
         }
